fix: add GetTimeOfDayForDex() to deal with pokedex mon areas

+++ include/pokedex_area_screen.h
@@ -1,7 +1,7 @@
 #ifndef GUARD_POKEDEX_AREA_SCREEN_H
 #define GUARD_POKEDEX_AREA_SCREEN_H
 
-extern u8 gAreaTimeOfDay;
+extern u32 gAreaTimeOfDay;
 
 enum ChangeTimeOfDay 
 {
